BF: DimArray now pickles, too, but with a possible hack.
So, I have made it so the dimension check for DimArray does not occur if the obj
passed into array_finalize is None.  This is what it is when you are unpickling and I
have not found any other case where it is None and all the tests still pass, so for
right now I think we are safe.